Bump to marshmallow 3 and fix up tests
