+++ tests/conftest.py
@@ -11,11 +11,11 @@ TEST_VALID_INPUT_YAML = join(TEST_DATA_ROOT_DIR, "valid_input.yaml")
 
 
 @pytest.fixture(scope="session")
-def base_valid_yaml():
+def base_valid_yaml_dict():
     with open(TEST_VALID_INPUT_YAML, "r") as stream:
         yield yaml.load(stream)
 
 
 @pytest.fixture(scope="function")
-def valid_yaml(base_valid_yaml):
-    yield deepcopy(base_valid_yaml)
+def valid_yaml_dict(base_valid_yaml_dict):
+    yield deepcopy(base_valid_yaml_dict)
